// src/interfaces/IClawSettings.ts
export interface IClawSettings {
	width: number
	height: number
	gravity: number
	friction: number
	groundFriction: number
	ballRadius: number
	clawSize: number
	clawWidth: number
	clawSpeedX: number
	clawSpeedY: number
	clawColor: string
	clawBoldColor: string
	dividerLineWidth: number
	dividerLineHeight: number
	dividerLineThickness: number
	dividerLineBorderColor: string
	dividerLineFillColor: string
}